requirements.txt
python-telegram-bot==20.7
supabase-py==2.3.1
python-dotenv==1.0.0